pyproject.toml
[tool.poetry]
name = "sql-usuario-post"
version = "0.1.0"
description = ""
authors = ["LivMad <livia.o.dantas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlmodel = "^0.0.18"
ruff = "^0.4.4"
ulid-py = "^1.1.0"
passlib = "^1.7.4"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
